+++ flaskr/routes/availabilities.py
@@ -1,13 +1,16 @@
 from flask import Blueprint, request, jsonify
-from flask_login import login_required
-from flaskr.models import db, Availability
 from datetime import datetime
+from flaskr.models import db, Availability
+from .utils import get_current_user_from_session, unauthorized_message
 
 bp = Blueprint('availabilities', __name__, url_prefix='/api/availabilities')
 
 @bp.route('/', methods=['POST'])
-@login_required
 def create_availability():
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     data = request.json
     new_availability = Availability(
         id=data['id'],
@@ -20,8 +23,11 @@ def create_availability():
     return jsonify({'message': 'Availability created'}), 201
 
 @bp.route('/<id>', methods=['PUT'])
-@login_required
 def update_availability(id):
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     data = request.json
     availability = Availability.query.get(id)
     if availability:
@@ -32,8 +38,11 @@ def update_availability(id):
     return jsonify({'message': 'Availability not found'}), 404
 
 @bp.route('/<id>', methods=['DELETE'])
-@login_required
 def delete_availability(id):
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     availability = Availability.query.get(id)
     if availability:
         db.session.delete(availability)
@@ -42,7 +51,10 @@ def delete_availability(id):
     return jsonify({'message': 'Availability not found'}), 404
 
 @bp.route('/', methods=['GET'])
-@login_required
 def list_availabilities():
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     availabilities = Availability.query.all()
     return jsonify([availability.serialize() for availability in availabilities]), 200
