+++ flaskr/routes/comments.py
@@ -1,13 +1,16 @@
 from flask import Blueprint, request, jsonify
-from flask_login import login_required
-from flaskr.models import db, Comment
 from datetime import datetime
+from flaskr.models import db, Comment
+from .utils import get_current_user_from_session, unauthorized_message
 
 bp = Blueprint('comments', __name__, url_prefix='/api/comments')
 
 @bp.route('/', methods=['POST'])
-@login_required
 def create_comment():
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     data = request.json
     new_comment = Comment(
         id=data['id'],
@@ -21,8 +24,11 @@ def create_comment():
     return jsonify({'message': 'Comment created'}), 201
 
 @bp.route('/<id>', methods=['PUT'])
-@login_required
 def update_comment(id):
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     data = request.json
     comment = Comment.query.get(id)
     if comment:
@@ -32,8 +38,11 @@ def update_comment(id):
     return jsonify({'message': 'Comment not found'}), 404
 
 @bp.route('/<id>', methods=['DELETE'])
-@login_required
 def delete_comment(id):
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     comment = Comment.query.get(id)
     if comment:
         db.session.delete(comment)
@@ -42,8 +51,11 @@ def delete_comment(id):
     return jsonify({'message': 'Comment not found'}), 404
 
 @bp.route('/', methods=['GET'])
-@login_required
 def list_comments():
+    user = get_current_user_from_session()
+    if not user:
+        return unauthorized_message()
+
     appointment_id = request.args.get('appointment_id')
     comments = Comment.query.filter_by(appointmentId=appointment_id).all()
     return jsonify([comment.serialize() for comment in comments]), 200
